Smush all classes together.
It really makes more sense this way.
